+++ Pods/Bolts/Bolts/Common/BFTask.m
@@ -12,8 +12,6 @@
 
 #import <libkern/OSAtomic.h>
 
-#import "BFExecutor.h"
-#import "BFTaskCompletionSource.h"
 #import "Bolts.h"
 
 __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
@@ -21,58 +19,68 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
           " Break on warnBlockingOperationOnMainThread() to debug.");
 }
 
+NSString *const BFTaskErrorDomain = @"bolts";
+NSString *const BFTaskMultipleExceptionsException = @"BFMultipleExceptionsException";
+
 @interface BFTask () {
     id _result;
     NSError *_error;
     NSException *_exception;
-    BOOL _cancelled;
 }
 
-@property (nonatomic, retain, readwrite) NSObject *lock;
-@property (nonatomic, retain, readwrite) NSCondition *condition;
-@property (nonatomic, assign, readwrite) BOOL completed;
-@property (nonatomic, retain, readwrite) NSMutableArray *callbacks;
+@property (atomic, assign, readwrite, getter = isCancelled) BOOL cancelled;
+@property (atomic, assign, readwrite, getter = isFaulted) BOOL faulted;
+@property (atomic, assign, readwrite, getter = isCompleted) BOOL completed;
+
+@property (nonatomic, strong) NSObject *lock;
+@property (nonatomic, strong) NSCondition *condition;
+@property (nonatomic, strong) NSMutableArray *callbacks;
+
 @end
 
 @implementation BFTask
 
-- (id)init {
+#pragma mark - Initializer
+
+- (instancetype)init {
     if (self = [super init]) {
-        self.lock = [[NSObject alloc] init];
-        self.condition = [[NSCondition alloc] init];
-        self.callbacks = [NSMutableArray array];
+        _lock = [[NSObject alloc] init];
+        _condition = [[NSCondition alloc] init];
+        _callbacks = [NSMutableArray array];
     }
     return self;
 }
 
-+ (BFTask *)taskWithResult:(id)result {
+#pragma mark - Task Class methods
+
++ (instancetype)taskWithResult:(id)result {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
     tcs.result = result;
     return tcs.task;
 }
 
-+ (BFTask *)taskWithError:(NSError *)error {
++ (instancetype)taskWithError:(NSError *)error {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
     tcs.error = error;
     return tcs.task;
 }
 
-+ (BFTask *)taskWithException:(NSException *)exception {
++ (instancetype)taskWithException:(NSException *)exception {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
     tcs.exception = exception;
     return tcs.task;
 }
 
-+ (BFTask *)cancelledTask {
++ (instancetype)cancelledTask {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
     [tcs cancel];
     return tcs.task;
 }
 
-+ (BFTask *)taskForCompletionOfAllTasks:(NSArray *)tasks {
++ (instancetype)taskForCompletionOfAllTasks:(NSArray *)tasks {
     __block int32_t total = (int32_t)tasks.count;
     if (total == 0) {
-        return [BFTask taskWithResult:nil];
+        return [self taskWithResult:nil];
     }
     
     __block int32_t cancelled = 0;
@@ -83,9 +91,7 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
     for (BFTask *task in tasks) {
         [task continueWithBlock:^id(BFTask *task) {
-            if (task.isCancelled) {
-                OSAtomicIncrement32(&cancelled);
-            } else if (task.exception) {
+            if (task.exception) {
                 @synchronized (lock) {
                     [exceptions addObject:task.exception];
                 }
@@ -93,30 +99,32 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
                 @synchronized (lock) {
                     [errors addObject:task.error];
                 }
+            } else if (task.cancelled) {
+                OSAtomicIncrement32(&cancelled);
             }
-            
+
             if (OSAtomicDecrement32(&total) == 0) {
-                if (cancelled > 0) {
-                    [tcs cancel];
-                } else if (exceptions.count > 0) {
+                if (exceptions.count > 0) {
                     if (exceptions.count == 1) {
-                        tcs.exception = [exceptions objectAtIndex:0];
+                        tcs.exception = [exceptions firstObject];
                     } else {
                         NSException *exception =
-                            [NSException exceptionWithName:@"BFMultipleExceptionsException"
-                                                    reason:@"There were multiple exceptions."
-                                                  userInfo:@{ @"exceptions": exceptions }];
+                        [NSException exceptionWithName:BFTaskMultipleExceptionsException
+                                                reason:@"There were multiple exceptions."
+                                              userInfo:@{ @"exceptions": exceptions }];
                         tcs.exception = exception;
                     }
                 } else if (errors.count > 0) {
                     if (errors.count == 1) {
-                        tcs.error = [errors objectAtIndex:0];
+                        tcs.error = [errors firstObject];
                     } else {
-                        NSError *error = [NSError errorWithDomain:@"bolts"
+                        NSError *error = [NSError errorWithDomain:BFTaskErrorDomain
                                                              code:kBFMultipleErrorsError
                                                          userInfo:@{ @"errors": errors }];
                         tcs.error = error;
                     }
+                } else if (cancelled > 0) {
+                    [tcs cancel];
                 } else {
                     tcs.result = nil;
                 }
@@ -127,7 +135,13 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     return tcs.task;
 }
 
-+ (BFTask *)taskWithDelay:(int)millis {
++ (instancetype)taskForCompletionOfAllTasksWithResults:(NSArray *)tasks {
+    return [[self taskForCompletionOfAllTasks:tasks] continueWithSuccessBlock:^id(BFTask *task) {
+        return [tasks valueForKey:@"result"];
+    }];
+}
+
++ (instancetype)taskWithDelay:(int)millis {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
     dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, millis * NSEC_PER_MSEC);
     dispatch_after(popTime, dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^(void){
@@ -135,6 +149,13 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     });
     return tcs.task;
 }
+
++ (instancetype)taskFromExecutor:(BFExecutor *)executor
+                       withBlock:(id (^)())block {
+    return [[self taskWithResult:nil] continueWithExecutor:executor withBlock:block];
+}
+
+#pragma mark - Custom Setters/Getters
 
 - (id)result {
     @synchronized (self.lock) {
@@ -180,6 +201,7 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
             return NO;
         }
         self.completed = YES;
+        self.faulted = YES;
         _error = error;
         [self runContinuations];
         return YES;
@@ -205,6 +227,7 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
             return NO;
         }
         self.completed = YES;
+        self.faulted = YES;
         _exception = exception;
         [self runContinuations];
         return YES;
@@ -214,6 +237,12 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
 - (BOOL)isCancelled {
     @synchronized (self.lock) {
         return _cancelled;
+    }
+}
+
+- (BOOL)isFaulted {
+    @synchronized (self.lock) {
+        return _faulted;
     }
 }
 
@@ -232,7 +261,7 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
             return NO;
         }
         self.completed = YES;
-        _cancelled = YES;
+        self.cancelled = YES;
         [self runContinuations];
         return YES;
     }
@@ -262,10 +291,12 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     }
 }
 
-- (BFTask *)continueWithExecutor:(BFExecutor *)executor
-                       withBlock:(BFContinuationBlock)block {
+#pragma mark - Chaining methods
+
+- (instancetype)continueWithExecutor:(BFExecutor *)executor
+                           withBlock:(BFContinuationBlock)block {
     BFTaskCompletionSource *tcs = [BFTaskCompletionSource taskCompletionSource];
-    
+
     // Capture all of the state that needs to used when the continuation is complete.
     void (^wrappedBlock)() = ^() {
         [executor execute:^{
@@ -278,7 +309,7 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
             }
             if ([result isKindOfClass:[BFTask class]]) {
                 [(BFTask *)result continueWithBlock:^id(BFTask *task) {
-                    if (task.isCancelled) {
+                    if (task.cancelled) {
                         [tcs cancel];
                     } else if (task.exception) {
                         tcs.exception = task.exception;
@@ -294,7 +325,7 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
             }
         }];
     };
-    
+
     BOOL completed;
     @synchronized (self.lock) {
         completed = self.completed;
@@ -305,18 +336,18 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     if (completed) {
         wrappedBlock();
     }
-    
+
     return tcs.task;
 }
 
-- (BFTask *)continueWithBlock:(BFContinuationBlock)block {
+- (instancetype)continueWithBlock:(BFContinuationBlock)block {
     return [self continueWithExecutor:[BFExecutor defaultExecutor] withBlock:block];
 }
 
-- (BFTask *)continueWithExecutor:(BFExecutor *)executor
-                withSuccessBlock:(BFContinuationBlock)block {
+- (instancetype)continueWithExecutor:(BFExecutor *)executor
+                    withSuccessBlock:(BFContinuationBlock)block {
     return [self continueWithExecutor:executor withBlock:^id(BFTask *task) {
-        if (task.error || task.exception || task.isCancelled) {
+        if (task.faulted || task.cancelled) {
             return task;
         } else {
             return block(task);
@@ -324,9 +355,11 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     }];
 }
 
-- (BFTask *)continueWithSuccessBlock:(BFContinuationBlock)block {
+- (instancetype)continueWithSuccessBlock:(BFContinuationBlock)block {
     return [self continueWithExecutor:[BFExecutor defaultExecutor] withSuccessBlock:block];
 }
+
+#pragma mark - Syncing Task (Avoid it)
 
 - (void)warnOperationOnMainThread {
     warnBlockingOperationOnMainThread();
@@ -338,13 +371,38 @@ __attribute__ ((noinline)) void warnBlockingOperationOnMainThread() {
     }
 
     @synchronized (self.lock) {
-        if (self.isCompleted) {
+        if (self.completed) {
             return;
         }
         [self.condition lock];
     }
     [self.condition wait];
     [self.condition unlock];
+}
+
+#pragma mark - NSObject
+
+- (NSString *)description {
+    // Acquire the data from the locked properties
+    BOOL completed;
+    BOOL cancelled;
+    BOOL faulted;
+
+    @synchronized (self.lock) {
+        completed = self.completed;
+        cancelled = self.cancelled;
+        faulted = self.faulted;
+    }
+
+    // Description string includes status information and, if available, the
+    // result sisnce in some ways this is what a promise actually "is".
+    return [NSString stringWithFormat:@"<%@: %p; completed = %@; cancelled = %@; faulted = %@;%@>",
+            NSStringFromClass([self class]),
+            self,
+            completed ? @"YES" : @"NO",
+            cancelled ? @"YES" : @"NO",
+            faulted ? @"YES" : @"NO",
+            completed ? [NSString stringWithFormat:@" result:%@", _result] : @""];
 }
 
 @end